+++ src/login.py
@@ -50,48 +50,3 @@ def parse_prof(cursor_table_desc, record):
         prof[field_name] = field_data
     return prof
 
-
-def input_password(prompt='Password: '):
-    print(prompt, end='', flush=True)
-    password = ''
-    
-    while True:
-        char = ""
-        if os.name == "nt":
-            char = msvcrt.getch().decode('utf-8')
-        else:
-            char = getch.getch()
-            
-        if char == '\r' or char == '\n':
-            break
-        
-        if char == '\x08':
-            password = password[:-1]
-            print('\b \b', end='', flush=True)
-        else:
-            password += char
-            print('*', end='', flush=True)
-    
-    print()
-    return password
-
-
-    # sys.stdout.write(prompt)
-    # sys.stdout.flush()
-    
-    # password = ''
-    # while True:
-    #     char = getch.getch()
-    #     if char == '\r' or char == '\n':
-    #         break
-    #     if char == '\x7f':
-    #         password = password[:-1]
-    #         sys.stdout.write('\b \b')
-    #         sys.stdout.flush()
-    #     else:
-    #         password += char
-    #         sys.stdout.write('*')
-    #         sys.stdout.flush()
-    # sys.stdout.write('\n')
-    # sys.stdout.flush()
-    # return password
