+++ src/ui.py
@@ -31,3 +31,52 @@ def print_menu(console, heading, tex_color,border_color, menu):
 #         .append("3. Options\n")
 #         .append("4. Exit program\n")
 # )
+
+
+
+def input_password(prompt='Password: '):
+    print(prompt, end='', flush=True)
+    password = ''
+    
+    while True:
+        char = ""
+        if os.name == "nt":
+            char = msvcrt.getch().decode('utf-8')
+        else:
+            char = getch.getch()
+            
+        if char == '\r' or char == '\n':
+            break
+        
+        
+        
+        if char == '\b':
+            password = password[:-1]
+            print('\b \b', end='', flush=True)
+        else:
+            password += char
+            print('*', end='', flush=True)
+    
+    print()
+    return password
+
+
+    # sys.stdout.write(prompt)
+    # sys.stdout.flush()
+    
+    # password = ''
+    # while True:
+    #     char = getch.getch()
+    #     if char == '\r' or char == '\n':
+    #         break
+    #     if char == '\x7f':
+    #         password = password[:-1]
+    #         sys.stdout.write('\b \b')
+    #         sys.stdout.flush()
+    #     else:
+    #         password += char
+    #         sys.stdout.write('*')
+    #         sys.stdout.flush()
+    # sys.stdout.write('\n')
+    # sys.stdout.flush()
+    # return password
